requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
h3==3.7.6
pytz==2023.3
pyyaml==6.0.1
python-json-logger==2.0.7
twilio==8.10.0
python-dotenv==1.0.0
schedule==1.2.0
pytest==7.4.0